WIP - Adapting native tests for iOS
testThaliMobileNative - Removed Android specific tests to their own file.
Started to adapt other tests to new peerAvailabilityChanged format and/or
to distinguish between how iOS and Android handle connect. Also extracted
a bunch of functions that are needed by this file as well as the Android
and iOS specific versions of this file.

testThaliMobileNativeAndroid - Moved native Android specific tests here.

testThaliMobileNativeiOS - Created a bunch of tests for basic iOS
functionality.

testUtils - Linting and investigating why Bluebird hates this file.

thaliMobileNativeTestUtils - Extracted a bunch of common code here from
testThaliMobileNative so it can be used by all 3 files

runTests - Useless attempts to placate Bluebird

thaliMobileNative.js - Spec changes! Check them out!
