+++ thali/NextGeneration/thaliMobileNative.js
@@ -440,7 +440,6 @@
  * | Illegal peerID | The peerID has a format that could not have been returned by the local platform |
  * | startListeningForAdvertisements is not active | Go start it! |
  * | Connection could not be established | The attempt to connect to the peerID failed. This could be because the peer is gone, no longer accepting connections or the radio stack is just horked. |
- * | Connection wait timed out | This is for the case where we are a lexically smaller peer and the lexically larger peer doesn't establish a connection within a reasonable period of time. |
  * | Max connections reached | The native layers have practical limits on how many connections they can handle at once. If that limit has been reached then this error is returned. The only action to take is to wait for an existing connection to be closed before retrying.  |
  * | No Native Non-TCP Support | There are no non-TCP radios on this platform. |
  * | No available TCP ports | There are no TCP ports available to listen on. |
@@ -451,7 +450,7 @@
  * @callback multiConnectResolvedCallback
  * @property {string} syncValue
  * @property {?string} error
- * @property {?number} port
+ * @property {?number} listeningPort
  */
 
 /**
@@ -468,10 +467,7 @@
 
 /**
  * Identifies the peerID of the peer with whom a `multiConnect` initiated
- * connection (read: MCSession) failed. This method MUST be fired when the
- * connection fails even if it is just because of a call to `disconnect`. If
- * this event is fired in direct response to a `disconnect` then error MUST be
- * null.
+ * connection (read: MCSession) failed.
  *
  * @public
  * @callback multiConnectConnectionFailureCallback
@@ -481,7 +477,12 @@
 
 /**
  * Fires the multiConnectConnectionFailureCallback if a multiConnect connection
- * fails for a reason other than a call to {@link disconnect}.
+ * fails. This failure can include a failure induced by a call to `disconnect`.
+ * Note, however, that this callback MUST only occur in response to an actual
+ * connection being terminated. So, for example, if disconnect is called with
+ * a peerID that isn't in the connected state then the disconnect will be
+ * successful but because no actual MCSession was terminated there won't be
+ * a multiConnectConnectionFailureCallback.
  *
  * @public
  * @function external:"Mobile(`multiConnectConnectionFailure`)".registerToNative
